added utils fule for discord bot
