+++ airflow/scripts/discord_scripts/discord_bot.py
@@ -3,6 +3,7 @@ from datetime import datetime, timedelta
 import discord
 from dotenv import load_dotenv
 import os
+from utils import *
 
 intents = discord.Intents.default()
 intents.members = True
@@ -33,9 +34,9 @@ async def on_message(message):
     if message.author.bot:
         return
     
-    update_member_activity(message)
+    update_member_activity(message, member_activity)
     if message.channel.category and is_projects_category(message.channel.category):
-        update_project_activity(message)
+        update_project_activity(message, project_activity)
 
     
 @client.event
@@ -63,38 +64,6 @@ async def reset_counters_weekly() -> None:
 
         print("Weekly counter reset " + datetime.now())
         await asyncio.sleep(7*24*60*60)
-
-
-def is_projects_category(category):
-    return category.name.lower().startswith('projects-')
-
-def is_text_channel(channel):
-    return channel.type.value == 0
-
-def update_member_activity(message):
-    author = message.author
-    channel = message.channel
-
-    if author.name not in member_activity:
-        member_activity[author.name] = {}
-    
-    if channel.name not in member_activity[author.name]:
-        member_activity[author.name][channel.name] = [0, 0]
-    
-    member_activity[author.name][channel.name][0] += 1
-    member_activity[author.name][channel.name][1] = message.created_at.strftime('%Y-%m-%d %H:%M:%S')
-    print(f'member activity: {member_activity}')
-
-def update_project_activity(message):
-    channel_name = message.channel.name
-    created_at = message.created_at
-
-    if channel_name not in project_activity:
-        project_activity[channel_name] = [0, 0]
-    
-    project_activity[channel_name][0] += 1
-    project_activity[channel_name][1] = created_at.strftime('%Y-%m-%d %H:%M:%S')
-    print(f'project activity: {project_activity}')
 
 
 client.run(BOT_TOKEN)
